--- a/lib/loader/loader.js
+++ b/lib/loader/loader.js
@@ -63,6 +63,7 @@ var loader, define, requireModule, require, requirejs;
     this.finalized = false;
     this.hasExportsAsDep = false;
     this.isAlias = alias;
+    this.reified = new Array(deps.length);
   }
 
   Module.prototype.makeDefaultExport = function() {
@@ -100,8 +101,8 @@ var loader, define, requireModule, require, requirejs;
   Module.prototype.reify = function() {
     var deps = this.deps;
     var length = deps.length;
-    var reified = new Array(length);
     var dep;
+    var reified = this.reified;
 
     for (var i = 0, l = length; i < l; i++) {
       dep = deps[i];